--- a/pilot/pkg/networking/core/v1alpha3/route/route.go
+++ b/pilot/pkg/networking/core/v1alpha3/route/route.go
@@ -97,12 +97,9 @@ type VirtualHostWrapper struct {
 // BuildSidecarVirtualHostsFromConfigAndRegistry creates virtual hosts from
 // the given set of virtual services and a list of services from the
 // service registry. Services are indexed by FQDN hostnames.
-func BuildSidecarVirtualHostsFromConfigAndRegistry(
-	node *model.Proxy,
-	push *model.PushContext,
-	serviceRegistry map[host.Name]*model.Service,
-	virtualServices []config.Config,
-	listenPort int) []VirtualHostWrapper {
+// The list of services is also passed to allow maintaining consistent ordering.
+func BuildSidecarVirtualHostsFromConfigAndRegistry(node *model.Proxy, push *model.PushContext, serviceRegistry map[host.Name]*model.Service,
+	services []*model.Service, virtualServices []config.Config, listenPort int) []VirtualHostWrapper {
 
 	out := make([]VirtualHostWrapper, 0)
 
@@ -117,9 +114,9 @@ func BuildSidecarVirtualHostsFromConfigAndRegistry(
 	}
 
 	// compute services missing virtual service configs
-	missing := make(map[host.Name]bool)
+	missing := make(map[host.Name]struct{})
 	for fqdn := range serviceRegistry {
-		missing[fqdn] = true
+		missing[fqdn] = struct{}{}
 	}
 	for _, wrapper := range out {
 		for _, service := range wrapper.Services {
@@ -128,8 +125,10 @@ func BuildSidecarVirtualHostsFromConfigAndRegistry(
 	}
 
 	// append default hosts for the service missing virtual services
-	for fqdn := range missing {
-		svc := serviceRegistry[fqdn]
+	for _, svc := range services {
+		if _, f := missing[svc.Hostname]; !f {
+			continue
+		}
 		for _, port := range svc.Ports {
 			if port.Protocol.IsHTTP() || util.IsProtocolSniffingEnabledForPort(port) {
 				cluster := model.BuildSubsetKey(model.TrafficDirectionOutbound, "", svc.Hostname, port.Port)