if we have domains like {foo, foo.local}, the order we build matters. If
foo is first, we will build the foo virtual host, then the foo.local.
But if foo.local is first, we build a single virtual host with both
domains, and then a second one with empty domain.

This is broken for three reasons
1. Empty domain list is rejected by envoy
2. The behavior is nondeterministic. Should have stable Service ordering
3. You may explicitly want the non-expanded versions