@howardjohn 2020-10-29
```
name                               old time/op        new time/op        delta
RouteGeneration/gateways-6               44.3ms ± 1%        44.7ms ± 1%   ~     (p=0.667 n=2+2)
RouteGeneration/gateways-shared-6         177ms ± 4%         179ms ± 4%   ~     (p=0.667 n=2+2)
RouteGeneration/empty-6                   760µs ± 1%        1357µs ± 0%   ~     (p=0.333 n=2+2)
RouteGeneration/tls-6                     768µs ± 2%        1379µs ± 1%   ~     (p=0.333 n=2+2)
RouteGeneration/telemetry-6               791µs ± 1%        1426µs ± 0%   ~     (p=0.333 n=2+2)
RouteGeneration/virtualservice-6         1.70ms ± 1%        1.73ms ± 2%   ~     (p=0.667 n=2+2)
```

benchmark shows pretty major increase, but for these cases the numbers in microseconds so its pretty much negligible?