@istio-testing 2020-10-29
@howardjohn: The following tests **failed**, say `/retest` to rerun all failed tests:

Test name | Commit | Details | Rerun command
--- | --- | --- | ---
gencheck_istio | db265a7361b0e25906c4476b2dbff73198b8b5db | [link](https://prow.istio.io/view/gs/istio-prow/pr-logs/pull/istio_istio/28414/gencheck_istio/1321931145423622144) | `/test gencheck_istio`
unit-tests_istio | db265a7361b0e25906c4476b2dbff73198b8b5db | [link](https://prow.istio.io/view/gs/istio-prow/pr-logs/pull/istio_istio/28414/unit-tests_istio/1321931141246095360) | `/test unit-tests_istio`

<details>

Instructions for interacting with me using PR comments are available [here](https://git.k8s.io/community/contributors/guide/pull-requests.md).  If you have questions or suggestions related to my behavior, please file an issue against the [kubernetes/test-infra](https://github.com/kubernetes/test-infra/issues/new?title=Prow%20issue:) repository. I understand the commands that are listed [here](https://go.k8s.io/bot-commands).
</details>
<!-- test report -->